Allow a mechanism to publish the split/merge request internally
Update the ApplicationProcessorEventsSource to be able to perform a
split and merge operation by publishing a SplitSegmentRequest and
MergeSegmentRequest on the ApplicationEventPublisher. The calls require
the calculation of respectively the biggest segment and smallest
segment, which should be done by using the ClientProcessors to deduce
the biggest/smallest segment for a given processor. Lastly, add some
javadoc, reindent the file according to the style guide and add a test
for the new functionality

#131