Fix split/merge in respect to several clients
The initial split/merge operations were send to every client in the
system which contained the given Tracking Event Processor.
That however means that a split will be performed by N processor, as
well as the merge - obviously incorrect.
This adjustment will ensure that the split is only performed by the
processor own the segment-to-split. Respectively, the merge is only
performed by the processor owning the segment-to-merge. Additionally,
all other clients are told to release the segment which is paired with
the segment-to-merge, to ensure nobody else can disturb the operation

#131