Use ClientProcessorsForContextAndName instead of Clients
The Clients iterator will make it so that any client receives a request
for status, event if it does not contain said Event Processor. Instead,
use the ClientProcessorsForContextAndName, which filters a given
ClientProcessors list based on the context and processor name provided
in the TEP to balance. Doing this ensure we will not incorrectly send
ProcessorStatusRequest. Additionally, perform some readability
adjustments

#support-216