Allow different mode types to enable streaming operations
The current implementation puts heavy emphasize on the `mode` field
present in the EventProcessor printable. Only when the mode equals
"Tracking" will operations like split and merge be available. To comply
with recent framework adjustments, anything tracking specific in this
area needs to be adjusting to Streaming. To that end, adding an
isStreaming field is beneficial, as this will mitigate the
equals-mode-validation entirely.

#enable-pooled-processor-mode