fix one more race condition
if first transaction claims a block in the event store, but before registers the writePosition in the Synchronizer a second transaction is completely written, the second transaction may be completed before the first one is.
This leads to an exception in an event processor, as it tries to read an uncompleted transaction. It could also lead to issues updating the index.