fix concurrency issue in notifyWritePositions (#496)
The iterator used to check the writePositions is only weakly consistent and may not see new entries in the underlying map.
If the current run has completed one or more writePositions it checks again for new entries.

if first transaction claims a block in the event store, but before registers the writePosition in the Synchronizer a second transaction is completely written, the second transaction may be completed before the first one is.
This leads to an exception in an event processor, as it tries to read an uncompleted transaction. It could also lead to issues updating the index.