fix missing events after transformation
when the last transaction in a segment contained multiple events and the last event was not updated in the transaction the last events were not copied into the new version of the segment.
Also, when a domain event was deleted, the event was no longer present in the index as the replacement event had no aggregate type.